Added a functional dvt-quorum method
In this patch, a quorum calculation feature was added. Not only is
quorum calculated, a list of developers who were counted is generated,
and the quorum and count are recorded. Also recorded are any missing
keys; that is, keys whose fingerprints live in the keyring but not in
LDAP. Also made minor, tested enhancements to gpg argument creation
code.

git-archimport-id: srivasta@debian.org--2003-primary/devotee--devel--0.1--patch-1