Enhancements
This patch updates dvt-gpg to be cognizant of the newer gpg versions
that pass in extra data if the sig was generated by a subkey.

Also, dvt-tally and dvt-voters have taken the lead from dvt-quorum,
and now record their output to a file. There are configuration options
to determine where the output shaould be placed. As always, a separate
dummy tally sheet is generated for use while voting is actually
underway. dvt-cron was changed to reflect the new behaviour of
dvt-tally and dvt-voters.



git-archimport-id: srivasta@debian.org--2003-primary/devotee--devel--0.1--patch-2