Added reults calculation
This release finishes off Devotee, adding the final piece. Cloneproof
SSD calculation of the winner. In conformance with the Debian
constitution, it gives due consideration to Debian's unique Quorum
rules, as well as Majority requirements. For that purpose, there is a
configuration option to se the Majority requirement for options whose
majority rewiurements differ fromt he defaulr 1:1 ratio.

The cron job was modified to also call the results calculator, and
cleaned up a trifle.

git-archimport-id: srivasta@debian.org--2003-primary/devotee--devel--0.1--patch-3