make vote acks MIME compliant, fix quorum output.
Many thanks to "Marco d'Itri" <md@linux.it> for determining that
making the ack for votes MIME compliant was not hard, and providing a
patch to dvt-ack. He motivated me to actually go read the RFC, and
massage the ack mail to be MIMEified.

AJ pointed to an erorr in dvt-quorum output, confusing K and Quorum
requirements. This has been fixed

git-archimport-id: srivasta@debian.org--2003-primary/devotee--devel--0.1--patch-20