Protect against spurious text being added at the end of a clear signed message in the body -- this way, an malicious attacker can not take an un related pgpg signed message, and tack on a spurious ballot to the end of that message, and send it to vote taking mechanism. Of course, even previously, when this attack would have succeeded, the ack would ahve been sent to the maintainer's email adress in LDAP, so the chance of it being undetected would have been low (well, the attacker could not place much confidence in the fact that the ballot would _not_ be detected as compromised).
Also protect against the possibility of someone sending in a ballot
for a previous vote and trying to pass it on as a ballot for the
curent vote. A full line check for each of the options in the ballot
was generating too many false negatives, so we now only check for the
leading word of each proposition -- and increase the weight of each
mismatch, so now each mismatch requires to be made up by at least
two matches.

Fix defeat comparison. The comparision is not by defeat margins, but
by the strength of the for fotes, or, in case of a tie, the smaller
against option wins. This required a reworking of the results
comparison segment (none of the votes run so far have been affected).

git-archimport-id: srivasta@debian.org--2003-primary/devotee--devel--0.1--patch-24