Fix generation of secret key for secret votes
Use /dev/urandom for the generated random number instead of using perl's rand
which is only seeded with a 32 bit random number.  Generate a longer key
(256 bit), stop using preditibaly data as part of the key, use sha256
instead of md5 and use an hmac function.

This is part of the fix for CVE-2012-2387

Based on suggestions from Raphael Geissert.